requirements.txt
fastapi
uvicorn
pandas
numpy
requests
beautifulsoup4
sentence-transformers
faiss-cpu
scikit-learn
streamlit
google-generativeai
python-dotenv
pydantic